Move game::handle_action
